fixed inode id size from 32bit to 64bit

+++ toolchain/udbfslib/set_boot_loader_inode.c
@@ -1,4 +1,4 @@
-// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/set_boot_loader_inode.c,v 1.1 2003/10/11 13:14:19 bitglue Exp $
+// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/set_boot_loader_inode.c,v 1.2 2003/10/12 18:05:35 instinc Exp $
 
 #define _LARGEFILE_SOURCE
 #define _LARGEFILE64_SOURCE
@@ -14,7 +14,7 @@
 
 int		udbfs_set_boot_loader_inode(
     UDBFSLIB_MOUNT	*mount,
-    uint32_t		inode_id ) {
+    uint64_t		inode_id ) {
 
   if( inode_id < mount->inode_count ) {
     mount->boot_loader_inode = inode_id;
